requirements.txt
albumentations==0.5.2
appdirs==1.4.4
catkin-pkg==0.4.23
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
distro==1.5.0
docutils==0.17.1
enum34==1.1.10
filterpy==1.4.5
h5py==2.10.0
imageio==2.9.0
imgaug==0.4.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.3.4
networkx==2.5.1
numba==0.53.1
numpy==1.19.5
opencv-python==4.4.0.46
opencv-python-headless==4.5.2.52
pandas==1.1.5
Pillow==8.0.1
plaidbench==0.7.0
plaidml==0.7.0
plaidml-keras==0.7.0
pycparser==2.20
pyopencl==2020.2.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytools==2021.2.6
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
rospkg==1.2.10
scikit-image==0.17.2
scipy==1.5.4
Shapely==1.7.1
six==1.16.0
tifffile==2020.9.3
tqdm==4.60.0
